refactor(): use repo/repoName instead of product/productName convention

+++ src/infrastructure/frameworks/getDiffInSeconds.ts
@@ -2,4 +2,4 @@
  * @description Get the difference in seconds between two moments in time (i.e. Unix timestamps).
  */
 export const getDiffInSeconds = (earlierTime: string, laterTime: string): number =>
-  Math.floor((parseFloat(laterTime) - parseFloat(earlierTime)) / 1000);
+  Math.floor((parseInt(laterTime) - parseInt(earlierTime)) / 1000);
